update spsa tuning params

+++ python_scripts/examples/apply_env_overrides.py
@@ -14,26 +14,26 @@ def make_tier_config(tier_lower, tier_upper):
             "g_lookahead_gac_local_min_unset", "g_lookahead_gac_local_max_unset", "g_lookahead_gac_global_min_entropy",
             "g_lookahead_constr_local_min_unset", "g_lookahead_constr_local_max_unset", "g_lookahead_constr_global_min_entropy"
         ],
-        "target_vars_template": """static const int	g_min_entropy_threshold = {g_min_entropy_threshold};
-static const int	g_gac_min_entropy = {g_gac_min_entropy};
-static const int	g_constr_min_entropy = {g_constr_min_entropy};
-static const double	g_lookahead_downgrade_fraction = {g_lookahead_downgrade_fraction};
+        "target_vars_template": """static const int		g_min_entropy_threshold = {g_min_entropy_threshold};
+static const int		g_gac_min_entropy = {g_gac_min_entropy};
+static const int		g_constr_min_entropy = {g_constr_min_entropy};
+static const double		g_lookahead_downgrade_fraction = {g_lookahead_downgrade_fraction};
 static const long long	g_period_scale = 1000000;
-static const int	g_period_coef_sqrt = {g_period_coef_sqrt};
-static const int	g_period_coef_inv = {g_period_coef_inv};
-static const int	g_period_coef_unset = {g_period_coef_unset};
-static const double	g_gac_local_min_unset = {g_gac_local_min_unset};
-static const double	g_gac_local_max_unset = {g_gac_local_max_unset};
-static const int	g_gac_global_min_entropy = {g_gac_global_min_entropy};
-static const double	g_constr_local_min_unset = {g_constr_local_min_unset};
-static const double	g_constr_local_max_unset = {g_constr_local_max_unset};
-static const int	g_constr_global_min_entropy = {g_constr_global_min_entropy};
-static const double	g_lookahead_gac_local_min_unset = {g_lookahead_gac_local_min_unset};
-static const double	g_lookahead_gac_local_max_unset = {g_lookahead_gac_local_max_unset};
-static const int	g_lookahead_gac_global_min_entropy = {g_lookahead_gac_global_min_entropy};
-static const double	g_lookahead_constr_local_min_unset = {g_lookahead_constr_local_min_unset};
-static const double	g_lookahead_constr_local_max_unset = {g_lookahead_constr_local_max_unset};
-static const int	g_lookahead_constr_global_min_entropy = {g_lookahead_constr_global_min_entropy};""",
+static const double		g_period_coef_sqrt = {g_period_coef_sqrt};
+static const double		g_period_coef_inv = {g_period_coef_inv};
+static const double		g_period_coef_unset = {g_period_coef_unset};
+static const double		g_gac_local_min_unset = {g_gac_local_min_unset};
+static const double		g_gac_local_max_unset = {g_gac_local_max_unset};
+static const int		g_gac_global_min_entropy = {g_gac_global_min_entropy};
+static const double		g_constr_local_min_unset = {g_constr_local_min_unset};
+static const double		g_constr_local_max_unset = {g_constr_local_max_unset};
+static const int		g_constr_global_min_entropy = {g_constr_global_min_entropy};
+static const double		g_lookahead_gac_local_min_unset = {g_lookahead_gac_local_min_unset};
+static const double		g_lookahead_gac_local_max_unset = {g_lookahead_gac_local_max_unset};
+static const int		g_lookahead_gac_global_min_entropy = {g_lookahead_gac_global_min_entropy};
+static const double		g_lookahead_constr_local_min_unset = {g_lookahead_constr_local_min_unset};
+static const double		g_lookahead_constr_local_max_unset = {g_lookahead_constr_local_max_unset};
+static const int		g_lookahead_constr_global_min_entropy = {g_lookahead_constr_global_min_entropy};""",
         "replacement_vars_template": f"""#include <stdlib.h>
 
 static int		g_min_entropy_threshold = {{g_min_entropy_threshold}};
@@ -41,9 +41,9 @@ static int		g_gac_min_entropy = {{g_gac_min_entropy}};
 static int		g_constr_min_entropy = {{g_constr_min_entropy}};
 static double	g_lookahead_downgrade_fraction = {{g_lookahead_downgrade_fraction}};
 static const long long	g_period_scale = 1000000;
-static int		g_period_coef_sqrt = {{g_period_coef_sqrt}};
-static int		g_period_coef_inv = {{g_period_coef_inv}};
-static int		g_period_coef_unset = {{g_period_coef_unset}};
+static double	g_period_coef_sqrt = {{g_period_coef_sqrt}};
+static double	g_period_coef_inv = {{g_period_coef_inv}};
+static double	g_period_coef_unset = {{g_period_coef_unset}};
 static double	g_gac_local_min_unset = {{g_gac_local_min_unset}};
 static double	g_gac_local_max_unset = {{g_gac_local_max_unset}};
 static int		g_gac_global_min_entropy = {{g_gac_global_min_entropy}};
@@ -79,13 +79,13 @@ static void	init_env(void)
 		g_lookahead_downgrade_fraction = atof(val);
 	val = getenv("{tier_upper}_PERIOD_COEF_SQRT");
 	if (val)
-		g_period_coef_sqrt = atoi(val);
+		g_period_coef_sqrt = atof(val);
 	val = getenv("{tier_upper}_PERIOD_COEF_INV");
 	if (val)
-		g_period_coef_inv = atoi(val);
+		g_period_coef_inv = atof(val);
 	val = getenv("{tier_upper}_PERIOD_COEF_UNSET");
 	if (val)
-		g_period_coef_unset = atoi(val);
+		g_period_coef_unset = atof(val);
 	val = getenv("{tier_upper}_GAC_LOCAL_MIN_UNSET");
 	if (val)
 		g_gac_local_min_unset = atof(val);
@@ -181,12 +181,12 @@ static void	init_routing_env(void)
         "var_names": [
             "g_sel_period_coef_sqrt", "g_sel_period_coef_inv"
         ],
-        "target_vars_template": """static const int			g_sel_period_coef_sqrt = {g_sel_period_coef_sqrt};
-static const int			g_sel_period_coef_inv = {g_sel_period_coef_inv};""",
+        "target_vars_template": """static const double			g_sel_period_coef_sqrt = {g_sel_period_coef_sqrt};
+static const double			g_sel_period_coef_inv = {g_sel_period_coef_inv};""",
         "replacement_vars_template": """#include <stdlib.h>
 
-static int			g_sel_period_coef_sqrt = {g_sel_period_coef_sqrt};
-static int			g_sel_period_coef_inv = {g_sel_period_coef_inv};
+static double			g_sel_period_coef_sqrt = {g_sel_period_coef_sqrt};
+static double			g_sel_period_coef_inv = {g_sel_period_coef_inv};
 
 #if !defined(G_PRUNE_NO_ENV) || !G_PRUNE_NO_ENV
 static void	init_sel_env(void)
@@ -198,10 +198,10 @@ static void	init_sel_env(void)
 		return ;
 	val = getenv("SEL_PERIOD_COEF_SQRT");
 	if (val)
-		g_sel_period_coef_sqrt = atoi(val);
+		g_sel_period_coef_sqrt = atof(val);
 	val = getenv("SEL_PERIOD_COEF_INV");
 	if (val)
-		g_sel_period_coef_inv = atoi(val);
+		g_sel_period_coef_inv = atof(val);
 	initialized = 1;
 }}
 #endif""",
